pyproject.toml
[project]
name = "chamados-projeto"
version = "0.1.0"
description = ""
authors = [
    {name = "Paulo Mateus",email = "paulo.fiden96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
